requirements.txt
pillow
pydantic~=2.9.0
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib